Take Playwright out of the CI-blocking build path
Cloudflare Workers Build failed again on the previous fix's commit.
Without log access to confirm the exact cause, the most likely one by
far: `prebuild` now launched Playwright's Chromium inside Cloudflare's
build container, which very commonly lacks the system libraries a full
browser needs — an unrelated, much more common failure mode than the
native resvg-js binary the previous commit fixed.

Splits the single generator into two independent scripts that were
being run together for no real reason:
- generate-static-images.mjs (Playwright): the 5 hand-authored section
  images. These don't depend on Sanity content, so there's nothing
  that goes stale between manual runs — no reason to pay for a browser
  launch on every build. Manual only, PNGs stay committed.
- generate-policy-images.mjs (satori + @resvg/resvg-js, no browser):
  one card per policy, tracking live content. This is the one with an
  actual reason to run automatically, so it's the only one `prebuild`
  invokes now.

Shared constants/helpers factored into lib.mjs, with no dependency on
playwright, satori, or @sanity/client — each generator only pulls in
what it actually needs.
